requirements.txt
Flask>=0.8
twilio>=3.3.6
Pillow>=3.3.0
zbarlight>=1.0
